requirements.txt
# Core LangChain + Community
langchain==0.1.0
langchain-community==0.0.10

# Embeddings and Vector DB
faiss-cpu==1.12.0
sentence-transformers==2.2.2

# Transformers stack
transformers==4.36.0
torch==2.4.1
accelerate==0.25.0
sentencepiece==0.1.99  # prebuilt wheels for Windows

# Speech / Whisper
openai-whisper==20231117
gTTS==2.4.0
pydub==0.25.1
SpeechRecognition==3.10.0

# Document Handling
pypdf==3.17.4
python-docx==1.1.0
pdfplumber==0.10.3

# Streamlit App
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
numpy>=1.25.0,<2
pandas==2.1.4

